pyproject.toml
[tool.poetry]
name = "django_ninja_auth"
version = "0.1.0"
description = "Django authorization views adapted to django-ninja"
authors = ["Martín Ugarte <contact@martinugarte.com>"]
license = "WTFPL"

[tool.poetry.dependencies]
python = "^3.6"
django-ninja = "^0.13.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
